Add some initial SSE4.2 support (#50)
This PR is based on #42 but contains a couple more function
implementation / fixes. With this patch, we are able to completely
render strips using SIMD (flattening works as well, thought there are
some problems with precision that I still need to debug). And while
there are still many functions missing, I think this is a good chunk of
implementation that can be merged first before doing further
implementation work,

---------

Co-authored-by: Benjamin Saunders <ben.e.saunders@gmail.com>
Co-authored-by: ajakubowicz-canva <ajakubowicz@canva.com>