Add more SSE4.2 impls (#52)
This is enough to make the whole test suite pass with SSE4.2! There is
still something wrong because performance is absolute trash on some
benchmarks (I presume some functions are not being inlined properly),
but at least the tests pass!

This also fixes an issue where we didn't use wrapping semantics in the
scalar code (this also applies to some other arithmetic operations, but
for now I've only fixed it for `mul`).