Handle a non-`Fallback` baseline (#105)
Fixes #74 

A couple of notes:
1) For ease, I removed `simd_dispatch` in this PR. As such, I don't want
to land this before [#simd >
v0.3.0](https://xi.zulipchat.com/#narrow/channel/514230-simd/topic/v0.2E3.2E0/with/543274721)
2) The cfg variants in here are gnarly; I believe that's unavoidable to
express what we want. We do need to be clear that we are doing runtime
dispatch properly
3) The automated testing story of this PR is an open question. I have a
manual test, in the new `sh ./check_targets.sh` at the root. I want to
defer running this in CI; I suspect it would be very expensive.

This does not address #92. This does however add a
`force_support_fallback` feature, so it doesn't completely remove the
fallback functionality.