Lazily document a rule of thumb for using Fearless SIMD (#110)
Ideally, we'd have a much fuller explanation of what's going on here.
But that's not currently forthcoming...

Thanks to Sergey "Shnatsel" Davidoff for motivating this in [#simd >
inlining](https://xi.zulipchat.com/#narrow/channel/514230-simd/topic/inlining/with/546913433).
I've added them as co-author based on their approval in [#simd >
inlining @
💬](https://xi.zulipchat.com/#narrow/channel/514230-simd/topic/inlining/near/546912914).

---------

Co-authored-by: "Sergey \"Shnatsel\" Davidoff" <shnatsel@gmail.com>