Build all tests for better cross-compilation (#126)
This was discussed a bit in #122. We were previously omitting
architecture-specific tests based on the *host* architecture, which is
incorrect and prevents cross-compiling tests to run under something like
qemu.

We can't access `CARGO_CFG_TARGET_ARCH` at the time that proc macros are
evaluated, so we just have to build every architecture-specific test and
use `#[cfg]` to disable ones that aren't applicable to the current
architecture.

Even more annoyingly, there's no way to conditionally ignore a test at
runtime. This means that if the CPU we're running the tests on doesn't
support a target feature, we have to either pass or fail that test. I've
chosen to fail, to prevent faulty code from appearing to pass tests.

I've also updated the architecture-specific `exclude` functions to add
`#[ignore]` attributes rather than omitting the tests entirely. This
makes it more clear that the tests exist, but are not being run because
they don't yet pass.