Implement all float<->int conversions on x86 (#134)
This PR implements fully-compliant conversions for f32->i32, f32->u32,
i32->f32, and u32->f32. They are implemented thusly:

- f32 to i32: Check if any operands are not less than 2147483648.0. This
catches out-of-range values and NaN, which is not less than any number.
If so, clamp the values and also handle NaN. This is about 25% slower
than raw `_mm_cvttps_epi32` in my microbenchmark.

- f32 to u32: Clamp any negative or NaN values to 0 using `_mm_max_ps`.
We then check again if any operands are above 2147483648.0. If so,
subtract that value from the input to give us the "residual" value,
convert it to i32, and add it to the final result. There is another
check to ensure we don't exceed the *unsigned* range. This is about as
fast as f32->i32.

- i32 to f32: This is easy, since it's the one intrinsic that Intel
actually gives us.

- u32 to f32: I will admit I don't fully understand how this one works.
It's what LLVM emits, and it's faster than [the other conversion I
found](https://stackoverflow.com/questions/34066228/how-to-perform-uint32-float-conversion-with-sse).
It requires no actual conversion instructions.

I've tested these conversions exhaustively on every 32-bit bit pattern,
and they all return the correct results.