Implement new min/max/min_precise/max_precise semantics (#136)
Resolves #133. Depends on #134, since I use a helper method I introduced
in that PR.

In #133, I brought up the question of what semantics we should use for
floating-point minimum and maximum operations. There were many different
options available, all with annoying tradeoffs on different
architectures.

~~I believe the conclusion we came to during office hours was to
implement the following:~~

- `min` and `max`:
  - If the first operand is NaN, return the second operand.
- If the second operand is NaN, return either the first operand or NaN
non-deterministically.
- Otherwise, if both operands are zero, regardless of sign, return
either non-deterministically.
  - Otherwise, return the minimum/maximum of the two operands.

These can be implemented using a single `minps`/`maxps`/`minpd`/`maxpd`
on x86 and the `pmin` and `pmax` operations in WebAssembly[^1], both of
which return the first operand if the second operand is NaN.

On AArch64 and (eventually) RISC-V, these can be implemented using
`fminnm`/`fmaxnm` (AArch64) and `fmin`/`fmax` (RISC-V), both of which
return the non-NaN operand if *either* the first or second operand is
NaN.

- `min_precise` and `max_precise`:
  - If the first operand is NaN, return the second operand.
  - If the second operand is NaN, return the first operand.
- Otherwise, if both operands are zero, regardless of sign, return
either non-deterministically.
  - Otherwise, return the minimum/maximum of the two operands.

On x86 and WebAssembly, these can be implemented using the looser
"return second operand if first is NaN" semantics, followed by a
check/select to pick the first operand if the second operand is NaN.

On AArch64 and RISC-V, these are implemented using the exact same
instructions as the regular `min` and `max`.

(EDIT: per a discussion on Zulip, we've decided to allow the non-precise
min/max to return NaN if either operand is NaN. This means some
algorithms might be forced to use the slower min_precise/max_precise on
x86 even if it's unnecessary, but could be faster on WebAssembly.)

---

As I noted in #133, there are some tradeoffs. For the chosen semantics,
WebAssembly draws the short straw.

WebAssembly exposes two different min/max operations. The `pmin`/`pmax`
operations return the first operand if the second is NaN[^1], and the
`min`/`max` operations return NaN if *either* operand is NaN.

`pmin` and `pmax` are efficient on x86, but require a compare/select on
AArch64 and probably RISC-V as well.

`min`/`max` are efficient on AArch64 (they map to the `fmin`/`fmax`
instructions). On x86, they're quite slow since their behavior needs to
be polyfilled, including signed-zero handling. They're probably also
slow on RISC-V, which doesn't have *any* instructions for "min/max, but
NaN if either operand is NaN".

WebAssembly's "relaxed SIMD" extension (now supported on all platforms)
provides "relaxed min/max" operations which non-deterministically
perform a `min`/`max` or a `pmin`/`pmax`. By requiring our non-precise
`min`/`max` to return a non-NaN value if the second operand is non-NaN,
we cannot use those relaxed operations.

If we instead chose *not* to guarantee semantics around NaN, and allowed
our non-precise `min` and `max` to return NaN if either operand is NaN,
we could make use of WebAssembly's "relaxed SIMD" operations. However,
this would lose performance in the case where client code knows that one
operand is non-NaN.

As I noted in #133, and as the Vello code motivating #23 demonstrates,
consumers of fearless_simd often know that one operand is non-NaN. In
that case, they can still use non-precise `min`/`max` and get efficient
code on both x86 and AArch64. If we loosened the non-precise `min`/`max`
semantics, however, they would be forced to use the precise versions,
costing an extra `cmpunord` and `blendv` on x86 that are completely
unnecessary.

As I mentioned, RISC-V doesn't expose *any* instructions directly
corresponding to WebAssembly's `min`/`max`. [There's an open issue about
adding such operations to
WebAssembly](https://github.com/WebAssembly/design/issues/1548), but
even if they were, the "relaxed SIMD" proposal has already been ratified
and couldn't make use of them.

Since WebAssembly has already locked itself into a design that will be
unavoidably slow on RISC-V, I'm willing to give them the short end of
the stick when it comes to performance.

[^1]: WebAssembly actually flips the operands so that it returns the
*second* operand if the *first* is NaN, but we can just flip them back.