Use WebAssembly's relaxed SIMD intrinsics if available (#143)
Resolves #142.

This PR makes use of WASM's new "relaxed SIMD" intrinsics if the
corresponding target feature is enabled. I opted against making it a new
SIMD level, since most operations are the exact same and WASM doesn't
have runtime feature detection anyway.

The relevant intrinsics for us are the relaxed fused multiply-add
(performs an FMA if the hardware supports it, or a regular multiply
followed by add if not), and relaxed "lane select" (selects based on the
high bit of each lane as x86 prefers, or bitwise as AArch64 prefers,
depending on the hardware). (EDIT: Following
https://github.com/linebender/fearless_simd/pull/136, we can also use
relaxed min/max operations for floats.)

It does mean we need to double our tests; maybe it would be worth adding
a `.cargo/config.toml` to remove some of the boilerplate?