Clean up method signature codegen (#156)
Instead of exporting functions from `OpSig` to generate the arguments
list and return type for each method separately, consolidate everything
into methods that return the entire method signature at once.

The only change to the generated code is in the parameter names of the
generic split/combine-based methods--they now match the parameter names
of the `Simd` trait methods.

I plan to add methods that use const generics as part of the swizzle
work, and this is a prerequisite.