Use native vector types to back SIMD types (#159)
Resolves https://github.com/linebender/fearless_simd/issues/150.

In #150, I began wondering if this would involve fully switching the
vector types over to being associated types on the `Simd` trait. For
trait coherence reasons, it turns out we don't want to do that. As such,
there *is* a meaningful reason to keep the generic vector types around,
and this change is much less intrusive than I'd previously thought.

This PR switches the SIMD vector types (`f32x4<S>`, etc) from storing
aligned arrays to storing architecture-specific vector types (e.g.
`__m128`, `float32x4_t`, `v128`). For vectors wider than the maximum
native vector width, arrays of vector types are used (e.g. `[__m128;
4]`). As I mentioned in #150, if we use arrays, we're essentially
relying on the autovectorizer to consistently turn array operations into
vector operations. The main point of writing a SIMD library is to not do
that. An example of this going wrong is [the workaround introduced to
prevent LLVM from pessimizing a vector load into a
`memset`](https://github.com/linebender/fearless_simd/blob/bcdc8b390689076be4349d64e4686e52e3b3c384/fearless_simd_gen/src/mk_simd_types.rs#L144-L149).
This workaround seems to have *prevented* some scalar Vello code from
being autovectorized.

To ensure that the arrays of native vector types are aligned to the
level they should be, they're wrapped in generic alignment-increasing
types (`Aligned128`, `Aligned256`, and `Aligned512`). I'm not sure if
this lets us remove the alignment attribute from the public-facing
vector types, so I've left it there.

The split/combine operations now directly operate on these wrapped
arrays instead of calling `copy_from_slice`. I've found that while the
slice copies could *usually* be elided, they were not *always* elided.
In addition, every single operation on wide vectors calls split/combine,
sometimes recursively. Requiring the compiler to optimize out all those
slice copies is both fragile and bad for compile times.

We do maintain the ability to treat the vector types as arrays. This
requires quite a bit of transmuting, which is unfortunate. It also
requires that the native vector types are in fact interpretable as
arrays. For the x86 types, the Rust documentation [guarantees
it](https://doc.rust-lang.org/core/arch/x86_64/struct.__m128.html),
saying this about all of them:

> The in-memory representation of this type is the same as the one of an
equivalent array (i.e. the in-memory order of elements is the same, and
there is no padding); however, the alignment is different and equal to
the size of the type. Note that the ABI for function calls may not be
the same.

The AArch64 and WASM documentation is not explicit about layout, but
their vector types *are* marked with `#[repr(simd)]`, which means the
same thing. It'd probably be good to get a guarantee that the layout is
part of the stable API surface, but I'm not sure who to ask. I've run
the test suite on Miri and it passes on x86, but can't run on AArch64 or
WebAssembly due to missing support for some intrinsics.

The operations for loading/storing vector types as arrays are now
methods on the `Simd` trait, implemented by each specific platform, just
like most other operations.

To avoid exposing the native vector types as part of the public API,
I've defined them as associated types on a new `ArchTypes` trait, which
is now a supertrait of `Simd`. The trait is not publicly exported. This
means I had to silence some `unnameable_types` lints on both `ArchTypes`
and the aligned wrapper types, which are also an internal implementation
detail.

I believe the only public-facing change is that the `val` field on
vector types is now private. You can still construct a vector type from
an array via `SimdFrom`/`SimdInto`, and convert a vector type to an
array (or a reference to one) via `Deref`/`DerefMut` or the
`From`/`Into` conversions. I kind of wonder if we should go all the way
and make the `simd` field private too, since you can access it via the
`witness` method.

If you want to test the performance impact of these changes, you can run
vello_bench on [this branch, without this
PR](https://github.com/valadaptive/vello/tree/fearless-simd-bench-pre-native-vectors),
and [this branch, which includes this
PR](https://github.com/valadaptive/vello/tree/fearless-simd-native-vectors).