Add `Level::is_fallback()` method (#168)
Implements some of what was discussed in
https://github.com/linebender/fearless_simd/issues/118.

In some cases, you might have a scalarized version of your algorithm
that runs faster than the `Fallback` SIMD version. It might be useful to
be able to fall back to it instead, by checking if SIMD is unsupported.

This is currently perilous, since `Level::Fallback` may not always be
present. This PR adds a method that lets you check if no SIMD level is
supported, even if `Level::Fallback` is absent.

The `#[cfg]` checks are getting a bit unwieldy. In the future, it may be
worth using something like
[`cfg_aliases`](https://github.com/katharostech/cfg_aliases), although
that would require introducing a `build.rs`.