Consolidate AVX2 and SSE4.2 generation into one module (#178)
This PR is stacked on top of #177.

After adding the `Level` trait as an abstraction, we can use it to merge
the AVX2 and SSE4.2 implementations into a single module. Most of the
logic is the same, except for a few functions which need to do something
different if AVX2 is supported. This gets rid of a lot of duplicate
code.